requirements.txt
fastapi
uvicorn[standard]
requests
tensorflow
torch
torchvision
Pillow
transformers
safetensors
numpy
python-multipart